// sictcwebClass-master/web/ux/angular/src/app/core/interfaces/Dealership.ts
export interface Dealership {
    RecordId: number;
    Name: string;
    DealershipId: number;
    DealerCode: string;
    OrganizationId: number;
    PrinterGroupId: number;
    Address: string;
    City: string;
    State: string;
    Zip: string;
    Phone: string;
    UserDealershipId: number;
    UsesAdvancedTags: boolean;
    TimeZone: string;
    IntegratedDealerId: number;
    IsIntegrated: boolean;
    Latitude: number;
    Longitude: number;
}